haciendo ej3

+++ src/ar/edu/ort/tp3/ej01/Vendedor.java
@@ -1,17 +1,17 @@
 package ar.edu.ort.tp3.ej01;
 
 public class Vendedor extends SubContratado {
-	private double porcentaje;
+	private float porcentaje;
 	
 	public Vendedor(String nombre, int edad, float cantHoras, float precioHora, float porcentaje) {
 		super(nombre, edad, cantHoras, precioHora);
-		this.porcentaje = porcentaje * 0.1;
+		this.porcentaje = porcentaje * 0.1f;
 	}
 	
 	@Override
 	public float calcularPago() 
 	{	
-		return (float) ((cantHoras * precioHora) * (0.1 * porcentaje + 1)) ;
+		return (cantHoras * precioHora) * (0.1f * porcentaje + 1) ;
 	}
 
 }
